requirements.txt
numpy>=1.26
opencv-python-headless>=4.10
scikit-learn>=1.5
joblib>=1.4
matplotlib>=3.9
pillow>=10.4
streamlit>=1.37
ipykernel>=6.29
